backend/requirements.txt
# Coirendevouz API — backend (Faz 1.1+)
# Python 3.12+ önerilir (3.14 ile test edildi)

Django==5.2.12
psycopg[binary]==3.3.3
python-dotenv==1.2.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0

# OpenAPI / şema (drf-spectacular)
attrs==26.1.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
uritemplate==4.2.0

# Django bağımlılıkları (pip freeze ile sabitlendi)
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2026.1
